requirements.txt
amqp==5.2.0
asgiref==3.7.2
attrs==23.2.0
az-iranian-bank-gateways==1.9.0
billiard==4.2.0
celery==5.3.6
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.9
django-cors-headers==4.3.1
django-filter==23.5
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
gunicorn==21.2.0
idna==3.6
inflection==0.5.1
isodate==0.6.1
itypes==1.2.0
Jinja2==3.1.3
kavenegar==1.1.2
kombu==5.3.5
lxml==5.1.0
Markdown==3.5.2
MarkupSafe==2.1.4
openapi-codec==1.3.2
packaging==23.2
pillow==10.2.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pycryptodome==3.20.0
PyJWT==2.8.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-file==2.0.0
requests-toolbelt==1.0.0
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.13
zeep==4.2.1
